pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-research-paper-generator"
version = "0.1.0"
description = "AI-powered research paper generator using LangChain agents"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "reportlab>=4.0",
    "chromadb>=0.5.0",
    "python-docx>=1.1.0",
    "PyPDF2>=3.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "langchain-text-splitters>=0.3.0",
    "duckduckgo-search>=6.0.0",
    "arxiv>=2.0.0",
    "python-dotenv>=1.0.0",
    "pymupdf>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "hypothesis>=6.100.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
